Links access fixed

+++ controllers/LinksController.php
@@ -64,18 +64,22 @@ class LinksController
 
     public function getBySearchAction()
     {
-        $query = $_POST['query'];;
-        $id = $_SESSION['user_id'];
+        if ($this->isAutorized()) {
+            $query = $_POST['query'];;
+            $id = $_SESSION['user_id'];
 
-        if ($this->isNotHtml($query)) {
-            if ($id == null) {
-                $this->view->redirect("users/signin");
+            if ($this->isNotHtml($query)) {
+                if ($id == null) {
+                    $this->view->redirect("users/signin");
+                } else {
+                    $links = $this->dbManager->Links->getBySearch($id, $query);
+                    $this->view->render("main", "links/getBySearch", $links);
+                }
             } else {
-                $links = $this->dbManager->Links->getBySearch($id, $query);
-                $this->view->render("main", "links/getBySearch", $links);
+                $this->view->redirect("links/getall");
             }
         } else {
-            $this->view->redirect("links/getall");
+            $this->view->redirect("users/signin");
         }
     }
 
@@ -93,27 +97,31 @@ class LinksController
 
     public function addNewAction()
     {
-        $userId = $_SESSION['user_id'];
-        $name = $_POST['name'];
-        $text = $_POST['text'];
-        $url = $_POST['url'];
+        if ($this->isAutorized()) {
+            $userId = $_SESSION['user_id'];
+            $name = $_POST['name'];
+            $text = $_POST['text'];
+            $url = $_POST['url'];
 
-        if ($this->isNotHtml($name) && $this->isNotHtml($text) && $this->isNotHtml($url)) {
-            $visibility = $_POST['visibility'];
-            if ($visibility == true) {
-                $visibility = 1;
-            } else {
-                $visibility = 0;
-            }
-            $now = new DateTime();
-            $date = $now->format('Y-m-d H:i:s');
+            if ($this->isNotHtml($name) && $this->isNotHtml($text) && $this->isNotHtml($url)) {
+                $visibility = $_POST['visibility'];
+                if ($visibility == true) {
+                    $visibility = 1;
+                } else {
+                    $visibility = 0;
+                }
+                $now = new DateTime();
+                $date = $now->format('Y-m-d H:i:s');
 
-            if ($name != "" && $url != "") {
-                if (filter_var($url, FILTER_VALIDATE_URL, FILTER_FLAG_PATH_REQUIRED)) {
+                if ($name != "" && $url != "") {
+                    if (filter_var($url, FILTER_VALIDATE_URL, FILTER_FLAG_PATH_REQUIRED)) {
 
-                    $this->dbManager->Links->addNew($userId, $name, $text, $url, $date, $visibility);
+                        $this->dbManager->Links->addNew($userId, $name, $text, $url, $date, $visibility);
 
-                    $this->view->redirect("links/getById");
+                        $this->view->redirect("links/getById");
+                    } else {
+                        $this->view->redirect("links/addForm");
+                    }
                 } else {
                     $this->view->redirect("links/addForm");
                 }
@@ -121,7 +129,7 @@ class LinksController
                 $this->view->redirect("links/addForm");
             }
         } else {
-            $this->view->redirect("links/addForm");
+            $this->view->redirect("users/signin");
         }
     }
     #endregion
@@ -142,28 +150,32 @@ class LinksController
 
     public function editAction()
     {
-        $rowId = $_POST['rowId'];
-        $userId = $_SESSION['user_id'];
-        $name = $_POST['name'];
-        $text = $_POST['text'];
-        $url = $_POST['url'];
-        $visibility = $_POST['visibility'];
+        if ($this->isAutorized()) {
+            $rowId = $_POST['rowId'];
+            $userId = $_SESSION['user_id'];
+            $name = $_POST['name'];
+            $text = $_POST['text'];
+            $url = $_POST['url'];
+            $visibility = $_POST['visibility'];
 
-        if ($this->isNotHtml($name) && $this->isNotHtml($text) && $this->isNotHtml($url)) {
-            if ($visibility == true) {
-                $visibility = 1;
-            } else {
-                $visibility = 0;
-            }
-            $now = new DateTime();
-            $date = $now->format('Y-m-d H:i:s');
+            if ($this->isNotHtml($name) && $this->isNotHtml($text) && $this->isNotHtml($url)) {
+                if ($visibility == true) {
+                    $visibility = 1;
+                } else {
+                    $visibility = 0;
+                }
+                $now = new DateTime();
+                $date = $now->format('Y-m-d H:i:s');
 
-            if ($name != "" && $url != "") {
-                if (filter_var($url, FILTER_VALIDATE_URL, FILTER_FLAG_PATH_REQUIRED)) {
+                if ($name != "" && $url != "") {
+                    if (filter_var($url, FILTER_VALIDATE_URL, FILTER_FLAG_PATH_REQUIRED)) {
 
-                    $this->dbManager->Links->edit($rowId, $userId, $name, $text, $url, $date, $visibility);
+                        $this->dbManager->Links->edit($rowId, $userId, $name, $text, $url, $date, $visibility);
 
-                    $this->view->redirect("links/getById");
+                        $this->view->redirect("links/getById");
+                    } else {
+                        $this->view->redirect("links/getById");
+                    }
                 } else {
                     $this->view->redirect("links/getById");
                 }
@@ -171,7 +183,7 @@ class LinksController
                 $this->view->redirect("links/getById");
             }
         } else {
-            $this->view->redirect("links/getById");
+            $this->view->redirect("users/signin");
         }
     }
     #endregion
@@ -179,11 +191,15 @@ class LinksController
     #region DELETE
     public function deleteByIdAction($post)
     {
-        $id = $post["id"];
+        if ($this->isAutorized()) {
+            $id = $post["id"];
 
-        $this->dbManager->Links->deleteById($id);
+            $this->dbManager->Links->deleteById($id);
 
-        $this->view->redirect("links/getById");
+            $this->view->redirect("links/getById");
+        } else {
+            $this->view->redirect("users/signin");
+        }
     }
     #endregion
 }
